requirements.txt
Flask==3.1.2
Flask-CORS==6.0.1
google-generativeai==0.8.3
gunicorn==21.2.0
